feat(actions/auth.action): Add dispatch type LOGOUT_SUCCESS and LOGOUT_FAILURE depending of the response api

+++ src/actions/auth.action.js
@@ -55,11 +55,19 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
    return async (dispatch) => {
+      dispatch({ type: authConstants.LOGOUT_REQUEST });
       const res = await axios.post("/admin/signout");
 
-      localStorage.clear();
-      dispatch({
-         type: authConstants.LOGOUT_REQUEST,
-      });
+      if (res.status === 200) {
+         localStorage.clear();
+         dispatch({
+            type: authConstants.LOGOUT_SUCCESS,
+         });
+      } else {
+         dispatch({
+            type: authConstants.LOGOUT_FAILURE,
+            payload: { error: res.data.error },
+         });
+      }
    };
 };
